+++ CHANGELOG.md
@@ -33,10 +33,15 @@ more easily in the future without breaking the public API of the client.
   documentation for more information on what each field represents. This allows customers apps to take action based on
   the telemetry (ex: adapting sampling rate based on the number of packets dropped). The telemetry sent to the agent
   hasn't changed so the same dashboard can be use for V4 and V5 apps.
+- Client side aggregation for Counts, Gauges and Sets is enabled by default. See `WithoutClientSideAggregation()` option
+  to disable it.
 
 ## Notes
 
 - [FEATURE] Adding public method `GetTelemetry` to retrieve the client internal telemetry since the start of the client.
+- [FEATURE] Client side aggregation for Counts, Gauges and Sets is enabled by default.
+  `WithExtendedClientSideAggregation()` for Timings, Histograms and Distributions is still disabled by default. Both
+  feature are no longer considered BETA.
 
 # 4.8.1 / 2021-07-09